refactor: update dependencies and clean up unused code in core and guest modules
