+++ winternitz/guest/Cargo.toml
@@ -10,9 +10,23 @@ risc0-zkvm = { version = "1.2.2", default-features = false, features = ['std'] }
 winternitz-core = { path = "../../core" }
 serde = "1.0"
 bitcoin = { version = "0.32.5", features = ["rand-std"] }
-tracing = { version = "0.1.40", default-features = false }
-tracing-subscriber = { version = "0.3.18", features = ["json"] }
+hex-literal = "0.4.1"
+num-bigint = "0.4.6"
+num-traits = "0.2.19"
+ark-ff = "0.5.0"
+ark-bn254 = "0.5.0"
+sha2 =  { version = "0.10.8", default-features = false }
+crypto-bigint = { version = "0.5.5", default-features = false }
+hex = "0.4.3"
+ark-groth16 = { version="0.5.0", default-features = false }
+ark-std = "0.5.0"
+ark-serialize = "0.5.0"
+ark-ec = "0.5.0"
 
 [[example]]
 name = "guest"
 path ="src/main.rs"
+
+[patch.crates-io]
+sha2 = { git = "https://github.com/risc0/RustCrypto-hashes", tag = "sha2-v0.10.8-risczero.0" }
+crypto-bigint = { git = "https://github.com/risc0/RustCrypto-crypto-bigint", tag = "v0.5.5-risczero.0" }
